+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Added
+- Added unit test coverage for check routers.
+
 ### Changed
 - The Backend struct has been refactored to allow easier customization for the
 enterprise edition.